feature/#47(WishService): Wish 예외처리

+++ BookHub/src/main/java/me/leeyeeun/bookhub/wish/service/WishService.java
@@ -2,6 +2,8 @@ package me.leeyeeun.bookhub.wish.service;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import me.leeyeeun.bookhub.global.exception.CustomException;
+import me.leeyeeun.bookhub.global.exception.Error;
 import me.leeyeeun.bookhub.member.entity.Member;
 import me.leeyeeun.bookhub.member.repository.MemberRepository;
 import me.leeyeeun.bookhub.wish.controller.dto.request.WishRequestDto;
@@ -30,13 +32,13 @@ public class WishService {
     @Transactional(readOnly = true)
     public Wish findWishById(Long id) {
         return wishRepository.findById(id)
-                .orElseThrow(() -> new IllegalArgumentException("위시글이 존재하지 않습니다."));
+                .orElseThrow(() -> new CustomException(Error.WISH_NOT_FOUND, Error.WISH_NOT_FOUND.getMessage()));
     }
 
     @Transactional(readOnly = true)
     public Member findMemberById(Long id) {
         return memberRepository.findById(id)
-                .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 회원입니다."));
+                .orElseThrow(() -> new CustomException(Error.NOT_FOUND_USER_EXCEPTION, Error.NOT_FOUND_USER_EXCEPTION.getMessage()));
     }
 
     @Transactional(readOnly = true)
@@ -65,7 +67,11 @@ public class WishService {
     @Transactional
     public Wish updateWish(Long wishId, WishRequestDto wishRequestDto, Principal principal) {
         Wish wish = findWishById(wishId);
-        getMemberFromPrincipal(principal);
+        Member member = getMemberFromPrincipal(principal);
+
+        if (!wish.getMember().getId().equals(member.getId())) {
+            throw new CustomException(Error.INVALID_USER_ACCESS, Error.INVALID_USER_ACCESS.getMessage());
+        }
 
         wish.update(wishRequestDto);
         return wish;
@@ -74,7 +80,11 @@ public class WishService {
     @Transactional
     public void deleteWish(Long wishId, Principal principal) {
         Wish wish = findWishById(wishId);
-        getMemberFromPrincipal(principal);
+        Member member = getMemberFromPrincipal(principal);
+
+        if (!wish.getMember().getId().equals(member.getId())) {
+            throw new CustomException(Error.INVALID_USER_ACCESS, Error.INVALID_USER_ACCESS.getMessage());
+        }
 
         wishRepository.delete(wish);
     }
@@ -89,8 +99,9 @@ public class WishService {
     public Wish findMyWishById(Long wishId, Principal principal) {
         Member member = getMemberFromPrincipal(principal);
         return wishRepository.findByIdAndMemberId(wishId, member.getId())
-                .orElseThrow(() -> new IllegalArgumentException("해당 위시글이 존재하지 않거나 권한이 없습니다."));
+                .orElseThrow(() -> new CustomException(Error.WISH_NOT_FOUND, "해당 독서록이 존재하지 않거나 회원에게 권한이 없습니다. id=" + wishId));
     }
+
     @Transactional(readOnly = true)
     public List<Wish> searchWishesByBookname(String keyword) {
         return wishRepository.findByBooknameContainingIgnoreCase(keyword);
